worker/requirements.txt
mysql-connector-python==9.2.0
pika==1.3.2
python-dotenv==1.0.1
pydantic==2.10.6